+++ src/views/Auction/index.tsx
@@ -24,6 +24,8 @@ import { Flex } from 'src/components/Flex'
 // Layouts
 import { Center } from 'src/layouts/Center'
 
+// Mesa Utils
+import { calculateClearingPrice } from 'src/mesa/price'
 // Views
 import { NotFoundView } from '../NotFound'
